requirements.txt
clickhouse_connect==0.8.18
fastapi==0.116.1
psycopg2_binary==2.9.10
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
sentence_transformers==5.0.0
uvicorn==0.35.0
